update: structure
