+++ src/product/product.controller.ts
@@ -1,6 +1,5 @@
 import { Product } from 'src/schemas/product.model'
 import { Body, Controller, Delete, Get, Param, Post, Put } from '@nestjs/common'
-import { MessagePattern } from '@nestjs/microservices'
 import { ApiTags } from '@nestjs/swagger'
 import { CreateDto, UpdateDto } from '../dtos/product.dto'
 import { ProductService } from './product.service'
@@ -11,25 +10,16 @@ export class ProductController {
   constructor(private productService: ProductService) {}
 
   @Get()
-  @MessagePattern({ name: 'all_product' })
   getAll() {
     return this.productService.getAll()
   }
 
-  @Get('main/:id')
+  @Get(':id')
   getOne(@Param('id') id: number) {
     return this.productService.getOne(id)
   }
 
-  // Used with redis microservice
-  @Get(':id')
-  @MessagePattern({ name: 'get_product' })
-  getOneProduct(id: number) {
-    return this.productService.getOne(id)
-  }
-
   @Post()
-  @MessagePattern({ name: 'product_created' })
   async addProduct(@Body() product: CreateDto) {
     const createDto = new Product()
     createDto.title = product.title
@@ -45,25 +35,8 @@ export class ProductController {
     return this.productService.update(id || product.id, product)
   }
 
-  // Microservice
-  @Put()
-  @MessagePattern({ name: 'product_updated' })
-  async productUpdatedRedis(payload: object) {
-    const updateDto = new UpdateDto()
-    updateDto.image = payload['image']
-    updateDto.title = payload['title']
-    return this.productService.update(payload['id'], updateDto)
-  }
-
   @Delete(':id')
   async productDeleted(@Param('id') id: number) {
-    const result = await this.productService.delete(id)
-    return result
-  }
-
-  @Delete()
-  @MessagePattern({ name: 'product_deleted' })
-  async productDeleteRedis(id: number) {
     const result = await this.productService.delete(id)
     return result
   }
