+++ src/product/product.module.ts
@@ -1,6 +1,7 @@
 import { Module } from '@nestjs/common'
 import { MongooseModule } from '@nestjs/mongoose'
 import { Product, ProductSchema } from '../schemas/product.model'
+import { ProductMicroController } from './product-microservice.controller'
 import { ProductController } from './product.controller'
 import { ProductService } from './product.service'
 
@@ -8,7 +9,7 @@ import { ProductService } from './product.service'
   imports: [
     MongooseModule.forFeature([{ name: Product.name, schema: ProductSchema }]),
   ],
-  controllers: [ProductController],
+  controllers: [ProductMicroController, ProductController],
   providers: [ProductService],
 })
 export class ProductModule {}
